requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
aiogram
python-dotenv
httpx